requirements.txt
python-docx==1.1.2
PyPDF2==3.0.1
langchain==0.3.2
langchain-community
langchain-huggingface
faiss-cpu
langchain_openai
langchain-groq